server/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
email-validator
